requirements.txt
yt-dlp
youtube-transcript-api
google-generativeai
flask
Flask-Cors
python-dotenv
markdown

